Do not set release information
Issue: release number across runners.
If a package is build on one runner the release starts at 1 and
increases. If the package is picked up by another runner it will
use the release state of this runner which can be 1 or >1 if
that package has been built by that runner in the past.
Thus the release number is non deterministic at the moment.
No good solution so far, don't want to interconnect services