Implement cb-collect service
cb-collect - fetches/updates a git repository and
collects build results of package sources as organized
in the git tree. Each project in the git tree will
be represented as a package repository. Currently only
the rpm format is detected and created. The implementation
is relatively easy and uses one thread per project to
create the repo and its metadata. The service will
constantly send package info requests to the message
broker using the same cb-collect consumer group. It's
load can currently only be controlled by the wait timeout
for inactivity.