Add response type setting to cb-info
Allow to configure the response behavior of cb-info. By
default a request to cb-info for a package/image for which the
runner has no information will result in no answer from
the service. If CB_INFO_RESPONSE_TYPE is set to

    --respond-always

the behavior changes in a way that cb-info will
send an answer for any received request, independently
if there is data available or not.