Fixed race condition in collector lock
Each project repo is created in its own thread. As long
as a thread works on a repo another attempt to create this
repo returns immediately with just a log information that
this repo setup is still in progress. For this purpose a
lock file is created. During the creation of the lock file
no other thread should be able to interfere with the file
locking code. This commit sets a mutex in the beginning
of _set_lock() and releases it at the end of the method
such that during runtime of _set_lock() no other thread
will call this method.