Update scheduler for clean buildroot
The scheduler was actively deleting the buildroot under certain
build flag conditions. This should not be done as there is now
the --clean option passed to the cb-run service which handles
the contents of the build root. Also the extra flags to handle
the rebuild_clean state were deleted. When building through a
request we always pass the --clean option to the cb-run service
to make sure those builds happen from a clean state. Only in
case of local builds the buildroot can be actively deleted or
kept.