Update documentation
Explain the from_scratch cluster setup a bit more generic
and avoid explicit aws calls. The chapter describing the
cluster setup in AWS should be used for this purpose. In
the from_scratch information I would like to put the focus
more on the setup steps which are no longer needed when
using the setup in AWS based on the pre-configured AMIs