Implement CBMessageBrokerSSHProxyKafka
Implementation of the message API to serve as SSH proxy.
There are the two tools ssh_kafka_read and ssh_kafka_write
which are used in an SSH connection to read/write data
to the kafka service running at the SSH endpoint. This is
a simply way to access kafka within the cloud if there is
no VPN or TLS access method available.