Refactor collecting build results
The first implementation uses a poll concept to ask each
runner for build results and fetch them. This creates a
high peak of messages and sync activity. The new concept
is based on a push model. Each runner pushes build results
to the collector and the collector creates repo metadata
on demand. This commit refactors the cb-scheduler, cb-run
and cb-image to be able to push their results to the
collector. The collector implementation itself has not yet
been rewritten and therefore the collector image is
currently not functional