Update documentation
An extra ebs volume for the collector is possible but imho
not strictly required since the root volume is also ebs its
size can be choosen at boot time. The advantage a resize of
this volume will auto resize on next boot which does not happen
on extra attached ebs volumes.

Also restructured the provisioning scripts a bit, such that
new runners can be added very easily