Prevent conflicting scheduler/info services
Scheduler and info services are called through the BlockingScheduler.
If the very first main loop exits, a blocking interval scheduler
enters the main loop again. In this condition the interval scheduler
would start new services after the timeout. This should not happen
because on the same runner(host) there should be only one active
scheduler and info service at a time.